Calculate line breaks for prose works (#9)
For dactylic hexameter, the line breaks specified in the source HTML are sufficient. But for prose we need to detect when a string of text is about to run off the page, and break the line accordingly. This PR adds that logic, as well as additional HTML cleanup. The tool now produces nice output for Catullus and Apuleius (Metamorphoses I)